% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ejemplo.R
\name{analiza}
\alias{analiza}
\alias{is.analizado}
\title{Ejemplo}
\usage{
analiza(datos, na.rm = FALSE)

is.analizado(x)
}
\arguments{
\item{datos}{Una matriz o data frame a ser analizado}

\item{na.rm}{Valor lógico que indica si los valores nulos deben
ser eliminados al obtener la información estadística descriptiva}

\item{x}{un objeto a comprobar}
}
\value{
Un objeto de clase \code{"analizado"}, básicamente una lista con
información estadística descriptiva
}
\description{
Un paquete de ejemplo en R.
}
\details{
Analiza una matriz y muestra información descriptiva de sus columnas.
No tiene ninguna finalidad práctica, mas que servir de recurso
didáctico para los programadores de R en lengua castellana.
Esta función permite obtener una vista general de una matriz
con información descriptiva tal como número de valores nulos,
máximo, mínimo, media, desviación estándar y varianza.
}
\examples{
#creamos una matriz 10x5
m <- matrix(runif(50), 10, 5)
m                    # objeto de clase 'matrix'
class(m)

#analizamos m
a <- analiza(m)
a                    # objeto de clase 'analizado'
class(a)

is.analizado(m)      # FALSE # m NO es un objeto de clase 'analizado'
is.analizado(a)      # TRUE  # a SÍ es un objeto de clase 'analizado'
}
\author{
José Antonio Muñoz Jiménez
}
\seealso{
\code{\link{summary}}
}

